Improved code, added new graph type (-N)  (#43)
The following changes have been made:

I've improved the README file by including command-line examples for every graph in the introduction.

The 2D bar chart could only show iodepth performance for a particular numjobs value. With the -N option, we can swap this and show numjobs performance for a particular iodepth.

There are some under-the-hood code improvements to make the code more consistent

There is some code that scaled down labels for x-axis if those become too wide otherwise. It is still wise to trim them using the --xlabel-parent and --xlabel-depth parameters if possible.